# FW: pre-allocate the stacks' threads