There are two reasons for this. First, it will allow us to randomise the stack addresses according to the fracture, replacing the code that was in `run_one_test()` and didn't do anything. Second, because it shows a performance improvement:

|  # Threads   | Before (µs)  |  After (µs)  |
|-------------:|-------------:|-------------:|
|            1 |        35950 |        33363 |
|           24 |        53527 |        53386 |
|           48 |        57227 |        55624 |
|           96 |        62519 |        59104 |
|          192 |        74021 |        67249 |

Subtracting the first line from all others:

<table>
<thead>
<tr>
<th rowspan=2># Threads</th>
<th colspan=2>Before (µs)</th>
<th colspan=2>After (µs)</th>
</tr>
<tr>
<th>Total</th>
<th>Per thread</th>
<th>Total</th>
<th>Per thread</th>
</tr>
</thead>
<tbody>
<tr>
<td align="right">23</td>
<td align="right">17576</td>
<td align="right">764</td>
<td align="right">20022</td>
<td align="right">871</td>
</tr>
<tr>
<td align="right">47</td>
<td align="right">21277</td>
<td align="right">453</td>
<td align="right">22261</td>
<td align="right">474</td>
</tr>
<tr>
<td align="right">95</td>
<td align="right">26568</td>
<td align="right">280</td>
<td align="right">25741</td>
<td align="right">271</td>
</tr>
<tr>
<td align="right">191</td>
<td align="right">38071</td>
<td align="right">199</td>
<td align="right">33886</td>
<td align="right">177</td>
</tr>
</tbody>
</table>

Even subtracting the overhead of starting a thread, the marginal cost of starting threads continues to decrease (I'm not sure why, I expected it to increase).

I also benchmarked not having guard pages, but that seems to trigger a perverse optimisation inside the kernel, at least in the machine I tested: it decided to offer 2 MB huge pages, but unfortunately that's slower to allocate.

Pre-faulting the top of the stack for each doesn't seem to help, because `pthread_create()` writes there anyway.